virtual_keyboard::add_testの仮引数の型がintだったのをdoubleに修正 close #30

+++ integration-sample/virtual-keyboard.cxx
@@ -24,7 +24,7 @@ namespace arisin
     void virtual_keyboard_t::reset()
     { pressing_keys_.clear(); }
     
-    void virtual_keyboard_t::add_test(const int x, const int y, const int stroke)
+    void virtual_keyboard_t::add_test(const double x, const double y, const double stroke)
     {
       L(INFO, "x(" << x << ") y(" << y << ") stroke(" << stroke << ")");
